add SCOPEv1 protocol
