+++ scopetools/templates/barcode_summary.html
@@ -22,7 +22,7 @@
             <p>Reads with lowQual：Barocde和UMI序列中包含大于1个碱基质量值低于14的Reads数目及比例</p>
             <p>Reads without linker：Read1序列中不包含linker的Reads数目及比例</p>
             <p>Reads without Barcode：Read1序列中不包含Barcode的Reads数目及比例</p>
-            <p>Reads with corrected Barcode: 校正Barcode的Reads数目</p>p>
+            <p>Reads with corrected Barcode: 校正Barcode的Reads数目</p>
         </div>
     </div>
 </div>
